Added a link finder from Canvas submissions
