+++ snap_file_getter.py
@@ -4,8 +4,8 @@ are easily downloadable from Canvas.
 """
 # TODO Automate this process as a part of running codediff on a folder of these html files.
 import argparse
-import logger
-from src.parser import HtmlParser
+import logging
+from src.scraper import scraperInstance
 from src import logger
 from src.utils import FileError, FileIOError, sdv
 
@@ -14,10 +14,7 @@ if __name__ == '__main__':
     argparser.add_argument('input_files', nargs='+',
                             help='input files for the file getter. Can be files and/or directories.')
     argparser.add_argument('-d', '--debug', help='Print debug statements for developers.',
-                            action='store_const', dest='log_level', const=logging.DEBUG)
-    verbose_group.add_argument('-q', '--quiet', help='Only output results.',
-                            action='store_const', dest='log_level', const=logging.ERROR)
-
+                        action='store_const', dest='log_level', const=logging.DEBUG)
     args = argparser.parse_args()
 
     _logger = logger.init_logger(args.log_level or logging.INFO)
@@ -25,8 +22,9 @@ if __name__ == '__main__':
     # running the program
     try:
         _logger.debug('Creating HtmlParser instance.')
-        html_parser = HtmlParser(args.input_files)
+        scraper = scraperInstance(args.input_files)
         _logger.debug('Scraping files.')
+
 
 
     except (FileError, FileIOError) as e:
